# swagger.yaml
openapi: 3.0.0
info:
  title: Health App API
  version: 1.0.0
  description: API for managing products and daily consumption

servers:
  - url: http://localhost:3000

paths:
  /products:
    get:
      summary: Search products
      description: Search for products by title using a query string
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: The search string
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /day/{date}:
    get:
      summary: Get day info
      parameters:
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD format
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Info for the specified day
          content:
            application/json:
              schema:
                type: object
                properties:
                  consumedProducts:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConsumedProduct'
                  totalKcal:
                    type: number

  /day/{date}/consumed:
    post:
      summary: Add consumed product
      parameters:
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                weight:
                  type: number
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Product added to the day

  /day/{date}/consumed/{productId}:
    delete:
      summary: Remove consumed product
      parameters:
        - in: path
          name: date
          required: true
          schema:
            type: string
        - in: path
          name: productId
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Product removed

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Product:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        kcal:
          type: number
        weight:
          type: number
        category:
          type: string

    ConsumedProduct:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'
        weight:
          type: number